project/requirements.txt
fastapi==0.95.1
uvicorn==0.23.1
python-dotenv==1.0.0
pillow==9.5.0
pytesseract==0.3.10
pandas==2.1.1
scikit-learn==1.2.2
joblib==1.2.0
numpy==1.24.3
